fix(cache): isolate runner telemetry failures

+++ packages/db/src/migrations/0000_mars_baseline.sql
@@ -211,13 +211,13 @@ CREATE TABLE IF NOT EXISTS worker_cache_status (
   active_snapshot_started_at timestamptz,
   last_completed_snapshot_id uuid,
   runner_cache_enabled boolean,
-  runner_cache_max_gib integer,
+  runner_cache_max_gib bigint,
   runner_cache_size_bytes bigint CHECK (runner_cache_size_bytes >= 0),
   runner_cache_entry_count bigint CHECK (runner_cache_entry_count >= 0),
   runner_cache_observed_at timestamptz
 );
 ALTER TABLE worker_cache_status ADD COLUMN IF NOT EXISTS runner_cache_enabled boolean;
-ALTER TABLE worker_cache_status ADD COLUMN IF NOT EXISTS runner_cache_max_gib integer;
+ALTER TABLE worker_cache_status ADD COLUMN IF NOT EXISTS runner_cache_max_gib bigint;
 ALTER TABLE worker_cache_status ADD COLUMN IF NOT EXISTS runner_cache_size_bytes bigint;
 ALTER TABLE worker_cache_status ADD COLUMN IF NOT EXISTS runner_cache_entry_count bigint;
 ALTER TABLE worker_cache_status ADD COLUMN IF NOT EXISTS runner_cache_observed_at timestamptz;
